+++ CHANGES.md
@@ -2,7 +2,9 @@
 
 ## unreleased
 
-- ✨ New Trigger: Bluetooth Device Connected
+- ✨ New Trigger: Bluetooth Device, launch commands on Bluetooth connection
+- ✨ BluetoothDeviceTrigger: Edit device by hand or from a list of paired or available
+- ✨ BluetoothDeviceTrigger: Show icon by major class and description by class
 
 ## 0.2.0 (2026-05-29)